refactor(server): move files to src/ dir
